Do not crash on unknown Ethernet packets
Fixed a problem that made the daemon crash when an unknown Ethernet frame was received and tcpdump printed the entire packet in hex over multiple lines. Since one line per packet was expected, the daemon crashed with a message like:

  File "/usr/local/sbin/macs.py", line 151, in run
    [time, mac] = self.process(line)
TypeError: 'NoneType' object is not iterable

By adding the -q option to tcpdump, only one line for such a packet will be generated.